Don't pause the game in singleplayer on the save schematics GUI
Since saving only happens when the game is not paused, having it pause on the save GUI creates an awkward pause when the save button is clicked but nothing happens.  Disabling pausing helps with this.

Perhaps pausing could be enabled for all UIs, but there's less of a benefit for the other ones, so I don't think it's needed there.